mall: 完善后台发送优惠券

+++ yudao-module-mall/yudao-module-promotion-biz/src/main/java/cn/iocoder/yudao/module/promotion/dal/dataobject/coupon/CouponDO.java
@@ -36,7 +36,7 @@ public class CouponDO extends BaseDO {
      *
      * 关联 {@link CouponTemplateDO#getId()}
      */
-    private Integer templateId;
+    private Long templateId;
     /**
      * 优惠劵名
      *
